+++ src/main/scala/co/topl/brambl/cli/impl/AssetStatementParserModule.scala
@@ -1,0 +1,9 @@
+package co.topl.brambl.cli.impl
+
+import cats.effect.IO
+
+trait AssetStatementParserModule {
+    def assetMintingStatementParserAlgebra(networkId: Int) =
+    AssetMintingStatementParser.make[IO](networkId)
+
+}