+++ src/main/scala/co/topl/brambl/cli/impl/SimpleMintingAlgebra.scala
@@ -6,7 +6,10 @@ import co.topl.brambl.builders.TransactionBuilderApi
 import co.topl.brambl.dataApi.GenusQueryAlgebra
 import co.topl.brambl.dataApi.WalletStateAlgebra
 import co.topl.brambl.models.Event
+import co.topl.brambl.models.box.AssetMintingStatement
 import co.topl.brambl.wallet.WalletApi
+import com.google.protobuf.ByteString
+import io.circe.Json
 
 trait SimpleMintingAlgebra[F[_]] {
   def createSimpleGroupMintingTransactionFromParams(
@@ -34,7 +37,7 @@ trait SimpleMintingAlgebra[F[_]] {
   ): F[Unit]
 
   def createSimpleAssetMintingTransactionFromParams(
-      keyFile: String,
+      keyfile: String,
       password: String,
       fromParty: String,
       fromContract: String,
@@ -42,6 +45,8 @@ trait SimpleMintingAlgebra[F[_]] {
       amount: Long,
       fee: Long,
       outputFile: String,
+      ephemeralMetadata: Option[Json],
+      commitment: Option[ByteString],
       assetMintingStatement: AssetMintingStatement
   ): F[Unit]
 }
@@ -60,7 +65,8 @@ object SimpleMintingAlgebra {
   ): SimpleMintingAlgebra[F] = new SimpleMintingAlgebra[F]
     with WalletApiHelpers[F]
     with GroupMintingOps[F]
-    with SeriesMintingOps[F] {
+    with SeriesMintingOps[F]
+    with AssetMintingOps[F] {
 
     override implicit val sync = psync
 
@@ -73,18 +79,13 @@ object SimpleMintingAlgebra {
     val wa = walletApi
 
     import co.topl.brambl.syntax._
-
-    override def createSimpleGroupMintingTransactionFromParams(
+    private def sharedOps(
         keyfile: String,
         password: String,
         fromParty: String,
         fromContract: String,
-        someFromState: Option[Int],
-        amount: Long,
-        fee: Long,
-        outputFile: String,
-        groupPolicy: Event.GroupPolicy
-    ): F[Unit] = for {
+        someFromState: Option[Int]
+    ) = for {
       keyPair <-
         walletManagementUtils
           .loadKeys(
@@ -103,8 +104,41 @@ object SimpleMintingAlgebra {
         fromParty,
         fromContract
       )
+    } yield (
+      keyPair,
+      predicateFundsToUnlock.get,
+      someCurrentIndices,
+      someNextIndices,
+      changeLock
+    )
+
+    override def createSimpleGroupMintingTransactionFromParams(
+        keyfile: String,
+        password: String,
+        fromParty: String,
+        fromContract: String,
+        someFromState: Option[Int],
+        amount: Long,
+        fee: Long,
+        outputFile: String,
+        groupPolicy: Event.GroupPolicy
+    ): F[Unit] = for {
+      tuple <- sharedOps(
+        keyfile,
+        password,
+        fromParty,
+        fromContract,
+        someFromState
+      )
+      (
+        keyPair,
+        predicateFundsToUnlock,
+        someCurrentIndices,
+        someNextIndices,
+        changeLock
+      ) = tuple
       fromAddress <- transactionBuilderApi.lockAddress(
-        predicateFundsToUnlock.get
+        predicateFundsToUnlock
       )
       response <- utxoAlgebra.queryUtxo(fromAddress)
       lvlTxos = response.filter(
@@ -112,7 +146,7 @@ object SimpleMintingAlgebra {
       )
       _ <- buildGroupTxAux(
         lvlTxos,
-        predicateFundsToUnlock.get.getPredicate,
+        predicateFundsToUnlock.getPredicate,
         amount,
         fee,
         someNextIndices,
@@ -137,26 +171,22 @@ object SimpleMintingAlgebra {
         outputFile: String,
         seriesPolicy: Event.SeriesPolicy
     ): F[Unit] = for {
-      keyPair <-
-        walletManagementUtils
-          .loadKeys(
-            keyfile,
-            password
-          )
-      someCurrentIndices <- getCurrentIndices(
+      tuple <- sharedOps(
+        keyfile,
+        password,
         fromParty,
         fromContract,
         someFromState
       )
-      predicateFundsToUnlock <- getPredicateFundsToUnlock(someCurrentIndices)
-      someNextIndices <- getNextIndices(fromParty, fromContract)
-      changeLock <- getChangeLockPredicate(
+      (
+        keyPair,
+        predicateFundsToUnlock,
+        someCurrentIndices,
         someNextIndices,
-        fromParty,
-        fromContract
-      )
+        changeLock
+      ) = tuple
       fromAddress <- transactionBuilderApi.lockAddress(
-        predicateFundsToUnlock.get
+        predicateFundsToUnlock
       )
       response <- utxoAlgebra.queryUtxo(fromAddress)
       lvlTxos = response.filter(
@@ -164,7 +194,7 @@ object SimpleMintingAlgebra {
       )
       _ <- buildSeriesTxAux(
         lvlTxos,
-        predicateFundsToUnlock.get.getPredicate,
+        predicateFundsToUnlock.getPredicate,
         amount,
         fee,
         someNextIndices,
@@ -183,7 +213,7 @@ object SimpleMintingAlgebra {
     } yield ()
 
     def createSimpleAssetMintingTransactionFromParams(
-        keyFile: String,
+        keyfile: String,
         password: String,
         fromParty: String,
         fromContract: String,
@@ -191,9 +221,74 @@ object SimpleMintingAlgebra {
         amount: Long,
         fee: Long,
         outputFile: String,
+        ephemeralMetadata: Option[Json],
+        commitment: Option[ByteString],
         assetMintingStatement: AssetMintingStatement
-    ): F[Unit] = ???
-
+    ): F[Unit] = for {
+      tuple <- sharedOps(
+        keyfile,
+        password,
+        fromParty,
+        fromContract,
+        someFromState
+      )
+      (
+        keyPair,
+        predicateFundsToUnlock,
+        someCurrentIndices,
+        someNextIndices,
+        changeLock
+      ) = tuple
+      fromAddress <- transactionBuilderApi.lockAddress(
+        predicateFundsToUnlock
+      )
+      response <- utxoAlgebra.queryUtxo(fromAddress)
+      lvlTxos = response.filter(
+        _.transactionOutput.value.value.isLvl
+      )
+      groupTxo <- response
+        .filter(
+          _.transactionOutput.value.value.isGroup
+        )
+        .find(_.outputAddress == assetMintingStatement.groupTokenUtxo)
+        .map(Sync[F].delay(_))
+        .getOrElse(
+          Sync[F].raiseError(
+            new Exception(
+              "Group token utxo not found"
+            )
+          )
+        )
+      seriesTxo <- response
+        .filter(
+          _.transactionOutput.value.value.isSeries
+        )
+        .find(_.outputAddress == assetMintingStatement.seriesTokenUtxo)
+        .map(Sync[F].delay(_))
+        .getOrElse(
+          Sync[F].raiseError(
+            new Exception(
+              "Group token utxo not found"
+            )
+          )
+        )
+      _ <- buildAssetTxAux(
+        keyPair,
+        outputFile,
+        lvlTxos,
+        predicateFundsToUnlock.getPredicate,
+        amount,
+        fee,
+        groupTxo.transactionOutput.value.value.group.get,
+        groupTxo.outputAddress,
+        someNextIndices,
+        seriesTxo.transactionOutput.value.value.series.get,
+        seriesTxo.outputAddress,
+        ephemeralMetadata,
+        commitment,
+        changeLock
+      )
+    } yield ()
   }
 
 }